requirements.txt
flask
transformers
pillow
torch
watchdict